added validator and modified services

+++ src/main/java/com/university/controller/DepartmentController.java
@@ -3,6 +3,7 @@ package com.university.controller;
 import com.university.collection.Departments;
 import com.university.entity.DepartmentEntity;
 import com.university.service.DepartmentService;
+import com.university.validator.DepartmentValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -18,6 +19,8 @@ import java.util.UUID;
 public class DepartmentController {
     @Autowired
     private DepartmentService service;
+    @Autowired
+    private DepartmentValidator departmentValidator;
     @GetMapping(path = "/test")
     public String test(){
         return "Welcome To Java";
@@ -44,12 +47,23 @@ public class DepartmentController {
     }
     @DeleteMapping(path = "/{id}")
     public ResponseEntity<Void> deleteDepartmentById(@PathVariable UUID id) {
-        service.deleteDepartmentById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        boolean isExist = departmentValidator.validateDepartmentExists(id);
+        if(isExist){
+            service.deleteDepartmentById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
     }
     @PutMapping(path = "/{id}")
     public ResponseEntity<Void> updateDepartment(@PathVariable UUID id, @RequestBody DepartmentEntity entity){
-        service.updateDepartment(entity);
-        return new ResponseEntity<Void>(HttpStatus.OK);
+        boolean isExist = departmentValidator.validateDepartmentExists(id);
+        if(isExist){
+            service.updateDepartment(entity);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 }
